requirements.txt
pytelegrambotapi
pixelmatch
pillow
numpy